requirements.txt
beautifulsoup4==4.10.0
boto3==1.21.33
botocore==1.24.33
certifi==2021.10.8
charset-normalizer==2.0.12
dnspython==2.2.1
idna==3.3
jmespath==1.0.0
marshmallow==3.15.0
packaging==21.3
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.9
